+++ grits/render.py
@@ -1,10 +1,11 @@
+import bs4
+from collections import ChainMap
 import jinja2
-import texas
 import shutil
-
-from contextlib import contextmanager
 from pathlib import Path
-from typing import Optional, Union
+from typing import MutableMapping, Mapping, Union
+
+DEFAULT_MAPP_OUTPUT_PATH = "/static/mapp.min.js"
 
 binary_suffixes = {
     "." + suffix
@@ -13,83 +14,108 @@ binary_suffixes = {
 }
 
 
-@contextmanager
-def temporary_value(d: dict, key: str, value):
-    """Temporarily set a key in a dict, then revert to previous value if it had one"""
-    missing = object()
-    tmp = d.get(key, missing)
-    d[key] = value
-    yield d
-    if tmp is missing:
-        del d[key]
-    else:
-        d[key] = tmp
-
-
 def default_is_binary(filename):
     """Used to tell the renderer which files to skip"""
     return Path(filename).suffix in binary_suffixes
 
 
+def safe_context(context: Mapping=None) -> ChainMap:
+    return ChainMap({}, (context or {}))
+
+
+def extract_head(soup: bs4.BeautifulSoup, context: MutableMapping) -> None:
+    head = None
+    head_sections = soup.find_all("head")
+    if head_sections:
+        if len(head_sections) > 1:
+            raise ValueError("html must have at most 1 <head> section.")
+        head = head_sections[0].renderContents().decode("utf-8")
+    context.setdefault("head", head)
+
+
+def extract_main(soup: bs4.BeautifulSoup, context: MutableMapping) -> None:
+    main_sections = soup.find_all("main")
+    if len(main_sections) != 1:
+        raise ValueError("html must have exactly 1 <main> section.")
+    main = main_sections[0]
+    context.setdefault("main", main.renderContents().decode("utf-8"))
+
+
+def prettify_html(rendered: str) -> str:
+    return bs4.BeautifulSoup(rendered, "html.parser").prettify()
+
+
+def extract_scripts(soup: bs4.BeautifulSoup, context: MutableMapping) -> None:
+    # Find top-level only, don't grab scripts in main twice
+    scripts = soup.find_all("script", recursive=False)
+    scripts = "\n".join((element.prettify()) for element in scripts)
+    context.setdefault("scripts", scripts)
+
+
 class Renderer:
-    def __init__(self, environment: jinja2.Environment, context: texas.Context, out_dir: Union[str, Path]):
+    def __init__(self, environment: jinja2.Environment, src_dir: Union[str, Path], out_dir: Union[str, Path]) -> None:
         self.environment = environment
-        self._context = context
-        self._context_views = ["default", "user"]
-        self._out_dir = Path(out_dir).expanduser()
+        self.src_dir = Path(src_dir).expanduser()
+        self.out_dir = Path(out_dir).expanduser()
+        self.out_dir.mkdir(parents=True, exist_ok=True)
 
-    @property
-    def out_dir(self):
-        self._out_dir.mkdir(parents=True, exist_ok=True)
-        return self._out_dir.resolve()
+    def render_scaffolding(self, context: Mapping=None) -> None:
+        """Required files to ensure mapp functions properly"""
+        context = safe_context(context)
+        self.render_template(name="_dynamicRoutes.json", context=context)
+        self.render_template(name="_prefetchManifest.json", context=context)
+        self.render_template(name="static/vendor/mapp.min.js", out_path="static/js/mapp.min.js", context=context)
 
-    @property
-    def context(self):
-        """read-only.  Modifications not preserved."""
-        return self._context.include(*self._context_views).snapshot
+    def render_template(self, name: str, out_path: Union[str, Path]=None, context: Mapping=None) -> None:
+        if out_path is None:
+            out_path = name
+        context = safe_context(context)
 
-    def render(self, name: str, dst_name: str=None, src_dir: Optional[Path]=None, context: dict=None):
-        """
-        render("style.css", "build/out.css") -> "build/out.css"
-        render("hello.html", "output/here.html") -> "output/here.html", "output/_/here.html"
-        """
-        context = context or self.context
-        if dst_name is None:
-            dst_name = name
+        template = self.environment.get_template(name)
+        rendered = template.render(context)
 
-        is_binary = context.get("is_binary", default_is_binary)
-        if is_binary(name):
-            name = str(src_dir / name)
-            self._render_binary(name, dst_name, context)
-        elif name.endswith(".html"):
-            with temporary_value(context, "page_filename", str(name)):
-                self._render_html(dst_name, context)
+        # fix templating ugliness
+        if name.endswith(".html"):
+            rendered = prettify_html(rendered)
+
+        dst = self.out_dir / out_path
+        dst.parent.mkdir(parents=True, exist_ok=True)
+        dst.write_text(rendered, encoding="utf-8")
+
+    def render_html(self, src_path: Union[str, Path], context: Mapping=None) -> None:
+        src_path = Path(src_path).expanduser()
+        out_path = src_path.relative_to(self.src_dir)
+
+        context = safe_context(context)
+        original = Path(src_path).expanduser().read_text(encoding="utf-8")
+        soup = bs4.BeautifulSoup(original, "html.parser")
+        try:
+            extract_head(soup, context)
+            extract_main(soup, context)
+            extract_scripts(soup, context)
+        except ValueError as error:
+            raise ValueError(f"While parsing '{str(src_path)}', encountered an error:") from error
+
+        self.render_template(name="__partial.html", out_path=Path("_") / out_path, context=context)
+        self.render_template(name="__full.html", out_path=out_path, context=context)
+
+    def render_asset(self, src_path: Union[str, Path], out_path: Union[str, Path]=None):
+        src_path = Path(src_path).expanduser()
+        if out_path:
+            out_path = Path(out_path).expanduser()
         else:
-            self._render_asset(name, dst_name, context)
-
-    def _render_html(self, dst_name: str, context: dict):
-        # 1. Render to /some/path/dst.html
-        tpl = self.environment.get_template("__full.html")
-        rendered = tpl.render(context)
-        dst = self.out_dir / dst_name
+            out_path = src_path.relative_to(self.src_dir)
+        dst = self.out_dir / out_path
         dst.parent.mkdir(parents=True, exist_ok=True)
-        dst.write_text(rendered, encoding="utf-8")
 
-        # 2. Render to /some/path/_/dst.html
-        tpl = self.environment.get_template("__partial.html")
-        rendered = tpl.render(context)
-        dst = self.out_dir / "_" / dst_name  # type: Path
-        dst.parent.mkdir(parents=True, exist_ok=True)
-        dst.write_text(rendered, encoding="utf-8")
+        shutil.copy(str(src_path), str(dst))
 
-    def _render_asset(self, name: str, dst_name: str, context):
-        tpl = self.environment.get_template(name)
-        rendered = tpl.render(context)
-        dst = self.out_dir / dst_name
-        dst.parent.mkdir(parents=True, exist_ok=True)
-        dst.write_text(rendered, encoding="utf-8")
-
-    def _render_binary(self, name: str, dst_name: str, context):
-        dst = self.out_dir / dst_name
-        dst.parent.mkdir(parents=True, exist_ok=True)
-        shutil.copy(name, str(dst))
+    def process(self, context: Mapping=None) -> None:
+        self.render_scaffolding(context=context)
+        for src_path in self.src_dir.glob("**/*"):
+            if src_path.is_dir():
+                continue
+            if src_path.suffix == ".html":
+                self.render_html(src_path=src_path, context=context)
+            else:
+                self.render_asset(src_path=src_path)